64 bit fixes

+++ logging.c
@@ -68,11 +68,11 @@ void log_request(struct wsgi_request *wsgi_req) {
 	if (uwsgi.shared->options[UWSGI_OPTION_MEMORY_DEBUG] == 1) {
 #ifndef UNBIT
 		rlen = snprintf(mempkt, 4096, "{address space usage: %lld bytes/%lluMB} {rss usage: %llu bytes/%lluMB} ",
-			uwsgi.workers[uwsgi.mywid].vsz_size, uwsgi.workers[uwsgi.mywid].vsz_size / 1024 / 1024,
-			uwsgi.workers[uwsgi.mywid].rss_size, uwsgi.workers[uwsgi.mywid].rss_size / 1024 / 1024);
+			(unsigned long long) uwsgi.workers[uwsgi.mywid].vsz_size, (unsigned long long ) uwsgi.workers[uwsgi.mywid].vsz_size / 1024 / 1024,
+			(unsigned long long) uwsgi.workers[uwsgi.mywid].rss_size, (unsigned long long ) uwsgi.workers[uwsgi.mywid].rss_size / 1024 / 1024);
 #else
 		rlen = snprintf(mempkt, 4096, "{address space usage: %lld bytes/%lluMB} ",
-			uwsgi.workers[uwsgi.mywid].vsz_size, uwsgi.workers[uwsgi.mywid].vsz_size / 1024 / 1024);
+			(unsigned long long) uwsgi.workers[uwsgi.mywid].vsz_size, (unsigned long long) uwsgi.workers[uwsgi.mywid].vsz_size / 1024 / 1024);
 #endif
 
 	
@@ -87,7 +87,7 @@ void log_request(struct wsgi_request *wsgi_req) {
 		uwsgi.mypid,
 		wsgi_req->app_id,
 		app_req,
-		uwsgi.workers[0].requests,
+		(unsigned long long ) uwsgi.workers[0].requests,
 		wsgi_req->remote_addr_len, wsgi_req->remote_addr,
 		wsgi_req->remote_user_len, wsgi_req->remote_user,
 		wsgi_req->var_cnt,
@@ -95,7 +95,7 @@ void log_request(struct wsgi_request *wsgi_req) {
 		24, time_request,
 		wsgi_req->method_len, wsgi_req->method,
 		wsgi_req->uri_len, wsgi_req->uri,
-		(uint64_t) wsgi_req->response_size,
+		(unsigned long long) wsgi_req->response_size,
 		(long int) (microseconds - microseconds2) / 1000,
 		via,
 		wsgi_req->protocol_len, wsgi_req->protocol,
